Fix HPZ Emerald state.
The type does not match the serialiser. This causes the game to
softlock during the LRZ3 Knuckles cutscene on the Wii U (and
presumably other platforms with different memory arrangements).

Correcting the serialiser instead of the type does not fix this
issue, so the opposite must be done.

+++ SonicMania/Objects/LRZ/HPZEmerald.h
@@ -19,7 +19,7 @@ struct ObjectHPZEmerald {
 struct EntityHPZEmerald {
     RSDK_ENTITY
     StateMachine(state); // unused
-    int32 type;
+    uint8 type;
     bool32 solid;
     Vector2 startPos;
     Hitbox *hitbox;
